Use DBDCAT instead of DBD.
Adding the generated DBD file to the DBD variable has the unintended
effect that it is generated with dbdExpand.pl. For older versions of
EPICS Base, this has the effect that empty record definitions are added
to the DBD file. For more recent versions, this leads to an error
because dbdExpand.pl complains about references to undefined records.