Avoid problems with scanIoRequest.
There seems to be a problem in EPICS Base (at least in version 7.0.6),
that sometimes the return value for a call to scanIoRequest indicates
success, but the record is not processed.

With the old implementation, this causes the record never to be updated
again while it remained in I/O Intr scan mode, because scanIoRequest
would not be called again by the routine handling subscription events
until a flag was cleared by the record processing routine.

The workaround that has been implemented is to always call scanIoRequest
when a subscription event arrives. The only downside of this approach is
that the same value might be used more than once if subscription events
arrive faster than the record can be processed. This is unlikely to
happen for any reasonable publishing interval.